Refactor: Move Note and NoteCardTestTags to library module
This commit moves the `Note` data class and the `NoteCardTestTags` object from the `com.mshdabiola.model` package to the `com.mshdabiola.library` package.

This change is reflected in the API dump files for Android, JVM, and KLIB, ensuring that the public API of the library module correctly exposes these types from their new location.

- Moved `com/mshdabiola/model/Note` to `com/mshdabiola/library/Note` in API dumps.
- Moved `com/mshdabiola/model/testtag/NoteCardTestTags` to `com/mshdabiola/library/NoteCardTestTags` in API dumps.
- Updated `NoteCard` function signature in API dumps to use `com/mshdabiola/library/Note`.