Refactor: Update KMP resources configuration
This commit refactors the project's build configuration to align with the new package naming for KMP resources. It updates the `compose.resources` block in the `library/build.gradle.kts` file by setting the `package` to `com.mshdabiola.library.resources`.

This change also resulted in the removal of the now-obsolete `kltemplate.library.generated.resources` package from the library's public API.